requirements.txt
black
click
kafka-python
mypy-extensions
pathspec
platformdirs
tomli
typing-extensions